Module Expr.Print

type 'a t = Format.formatter -> 'a -> unit

Alias for the type printing functions.

val print_tags : bool ref

Determine whether to print the map of tags for each id/type/term or not.

val print_index : bool ref

Determines whether to print the unique index of each identifier or not.

val name : Pretty.name Tag.t

The name tag is used for the printing of identifiers. When an identifier has an name tag, its value is used instead of the identifier intrinsic name.

val pos : Pretty.pos Tag.t

Positioning for pretty printing. If this tag is set, the printing functions will ignore type arguments (for readability). Pretty.Infix uses the identifier as a separator when printing th argument list Pretty.Prefix just ignore type arguments.

val id : _ id t

Printer for ids

val type_ : type_ t

Printer for type_.

val type_fun : type_fun t

Printer for type_fun.

val ty_var : ty_var t

Printer to print type variables along with their types.

val ty_cst : ty_cst t

Printer to print type constants along with their types.

val term_var : term_var t

Printer to print term variables along with their types.

val term_cst : term_cst t

Printer to print term constants along with their types.

val ty : ty t

Printer for types.

val term : term t

Printer for terms.

val formula : formula t

Printer for formulas.

val ty_def : ty_def t

Printer for type definitions.