Module Const.String

A module for string constant symbols that occur in terms.

val string : string -> t

String literal.

val length : t

String length.

val at : t

Get a char in a string.

val to_code : t

Returns the code point of a the single character of the string or (-1) if the string is not a singleton.

val of_code : t

Returns the singleton string whose only character is the given code point.

val is_digit : t

Check if the string a singleton string with a single digit character.

val to_int : t

Evaluates the string as a decimal natural number, or (-1) if it's not possible.

val of_int : t

Convert an int expression to a string in decimal representation.

val concat : t

String concatenation.

val sub : t

Substring extraction.

val index_of : t

Index of the first occurrence of the second string in first one, starting at the position of the third argument.

val replace : t

Replace the first occurrence.

val replace_all : t

Replace all occurrences.

val replace_re : t

Replace the leftmost, shortest re ocurrence.

val replace_re_all : t

Replace left-to-right, each shortest non empty re occurrence.

val is_prefix : t

Check if the first string is a prefix of the second one.

val is_suffix : t

Check if the first string is a suffix of the second one.

val contains : t

Check if the first string contains the second one.

val lt : t

Check for lexicographic strict ordering.

val leq : t

Check for lexicographic large ordering.

val in_re : t

Check if the string is in regular language.

module Reg_Lang : sig ... end

A module for regular language constant symbols that occur in terms.