Parameter Make.Parse

Main interface

exception Error of int

Exception raised by the parser when it encounters an error.

val file : (Lexing.lexbuf -> Ty.token) -> Lexing.lexbuf -> Ty.statement list

A function that parses an entire file, i.e until the end-of-file token, and return the list of parsed statements.

val input : (Lexing.lexbuf -> Ty.token) -> Lexing.lexbuf -> Ty.statement option

A function to parse a single statement. Returns None if it encounters the end-of-file token. Used for incremental parsing.